Move config.Store's new methods the right file
- rename the receivers so they are all `s` (they were intermingled even
  before the merge)
